Add custom add-to-cart dialog
