--- src/themes/icmaa-imp/components/core/AddToCartDialog.vue
@@ -1,16 +1,0 @@
-<template>
-  <sidebar>
-    Hi
-  </sidebar>
-</template>
-
-<script>
-import Sidebar from 'theme/components/theme/blocks/AsyncSidebar/Sidebar'
-
-export default {
-  name: 'AddToCartDialog',
-  components: {
-    Sidebar
-  }
-}
-</script>
